adding check version control

+++ installation.sh
@@ -1,9 +1,15 @@
 #!/bin/bash
-
 # Exit on error
 set -e
+# Checking Ubuntu version
+echo "Checking Ubuntu version..."
+if ! grep -q 'ID=ubuntu' /etc/os-release || ! grep -q 'VERSION_ID="24.04"' /etc/os-release; then
+  echo "Error: This script works only on Ubuntu 24.04."
+  exit 1
+fi
+echo "Ubuntu 24.04 detected. Continuing with the installation..."
 
-# Step 1: Uninstalling conflicting packages
+#Uninstalling conflicting packages
 echo "Removing conflicting packages..."
 for pkg in docker.io docker-doc docker-compose docker-compose-v2 podman-docker containerd runc; do
   sudo apt-get remove -y $pkg || true
@@ -13,14 +19,13 @@ echo "Installing dependencies..."
 sudo apt-get update
 sudo apt-get install -y ca-certificates curl
 
-# Step 2: Adding Docker GPG key and repository. The commands are copied from the official Docker installation guide
+# Adding Docker GPG key and repository. The commands are copied from the official Docker installation guide
 echo "Adding Docker GPG key and repository..."
 sudo install -m 0755 -d /etc/apt/keyrings
 sudo curl -fsSL https://download.docker.com/linux/ubuntu/gpg -o /etc/apt/keyrings/docker.asc
 sudo chmod a+r /etc/apt/keyrings/docker.asc
 echo "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.asc] https://download.docker.com/linux/ubuntu \n$(. /etc/os-release && echo "$VERSION_CODENAME") stable" | sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
-
-# Step 3: Installing Docker
+# Installing Docker
 echo "Installing Docker..."
 sudo apt-get update
 sudo apt-get install -y docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin
@@ -29,10 +34,11 @@ sudo apt-get install -y docker-ce docker-ce-cli containerd.io docker-buildx-plug
 echo "Verifying Docker installation..."
 sudo docker run hello-world
 
-# Step 4: Cloning the project repository
+# Cloning the project repository
 echo "Cloning project repository..."
 
-PROJECT_DIR=$(basename "$REPO_URL" .git)
+PROJECT_DIR="Linux-network-management--Docker--Add-Prodicts"
+
 if [ -d "$PROJECT_DIR" ]; then
   echo "Project directory already exists. Pulling latest changes..."
   cd "$PROJECT_DIR"
@@ -42,7 +48,7 @@ else
   cd "$PROJECT_DIR"
 fi
 
-# Step 5: Runing the project with Docker Compose
+# Runing the project with Docker Compose
 echo "Starting the project with Docker Compose..."
 sudo docker-compose up -d
